% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_for_nested_missing.R
\name{account_for_nested_missing}
\alias{account_for_nested_missing}
\title{Account for Data that is Missing Due to Nesting}
\usage{
account_for_nested_missing(df)
}
\arguments{
\item{df}{A dataframe containing cleaned ETS data as produced by \code{tbinenglanddataclean}.}
}
\value{
A dataframe that has nested missingness corrected for a subset of variables
}
\description{
This function adds entries for variables that are only missing due to nesting - rather
than truely being missing. Variables that are adjusted are: cause of death, date of death, starting treatment date,
treatment end date, year of BCG vaccination, and time since entry.
}
\examples{


## Code
account_for_nested_missing
}
