% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_results_folder.R
\name{make_results_folder}
\alias{make_results_folder}
\title{Make a Results Folder}
\usage{
make_results_folder(prefix, regen = FALSE)
}
\arguments{
\item{prefix}{A character string containing the name of the results folder to create}

\item{regen}{Logical - defaulting to \code{FALSE}. Should the results folders be regenerated.}
}
\value{
A path to the results folder.
}
\description{
This utility function makes a results folder with the specified name within a results folder. It will
also overwrite old results if required. Returns a path to the results folder.
}
\examples{

## Code
make_results_folder
}
