% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_variable_names.R
\name{map_variable_names}
\alias{map_variable_names}
\title{Maps Coded Variable Names to Human Readable Names}
\usage{
map_variable_names(df)
}
\arguments{
\item{df}{A dataframe with a variable field.}
}
\value{
A datafame with human readable variable entries.
}
\description{
This function - specific to this analysis - changes coded
variable names to be human readable using a predefined look-up.
}
\examples{

## Code
map_variable_names
}
