% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_missing_table.R
\name{nested_missing_table}
\alias{nested_missing_table}
\title{Tabulate Nested Missing Data}
\usage{
nested_missing_table(df)
}
\arguments{
\item{df}{A dataframe as produced by \code{\link[ETSMissing]{account_for_nested_missing}}}
}
\value{
A dataframe summarising missing data
}
\description{
This function tabulates nested missing data using a slightly altered (for presentation) version of
\code{\link[naniar]{miss_var_summary}}.
}
\examples{

## Code
nested_missing_table
}
