% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nested_missing.R
\name{plot_nested_missing}
\alias{plot_nested_missing}
\title{Plot Nested Missing Data}
\usage{
plot_nested_missing(df, down_sample = 0.2)
}
\arguments{
\item{df}{A dataframe as produced by \code{\link[ETSMissing]{account_for_nested_missing}}}

\item{down_sample}{Numeric, the fraction by which to downsample the data. Defaults to 0.2.}
}
\value{
A plot of nested missingness
}
\description{
This function plots nested missing data using a slightly altered version of
\code{\link[naniar]{vis_miss}}.
}
\examples{

## Code
plot_nested_missing
}
\seealso{
account_for_nested_missing
}
