% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull_results}
\alias{pull_results}
\title{Pull out Results by Variable}
\usage{
pull_results(df, var)
}
\arguments{
\item{df}{A dataframe containing an outcome variable (character) and a missing_table variable.}

\item{var}{A character string identifying which variable to pull results for.}
}
\value{
A dataframe of results
}
\description{
Pull out Results by Variable
}
\examples{

## Code
pull_results
}
