% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_data}
\alias{read_data}
\title{Read in results}
\usage{
read_data(name, path = NULL)
}
\arguments{
\item{name}{Character string identifying the data}

\item{path}{Character string giving the folder path}
}
\value{
The saved object
}
\description{
Read in results
}
\examples{

##Code
read_data
}
