% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_data}
\alias{save_data}
\title{Save results}
\usage{
save_data(df, name, path = NULL)
}
\arguments{
\item{df}{An object}

\item{name}{Character string to use to save the object}

\item{path}{Character string giving the folder path}
}
\value{
Nothing returned
}
\description{
Save results
}
\examples{

##Code
save_data
}
