% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_figure}
\alias{save_figure}
\title{Save Figures a PNGs}
\usage{
save_figure(fig, name, path = NULL)
}
\arguments{
\item{fig}{\code{ggplot2} object.}

\item{name}{A character string giving the name for the figure}

\item{path}{Character string giving the folder path}
}
\value{
Nothing returned
}
\description{
Save Figures a PNGs
}
\examples{

## Code
save_figure
}
