% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_figure}
\alias{show_figure}
\title{Load and Show Saved Figures}
\usage{
show_figure(name, path = NULL)
}
\arguments{
\item{name}{Character string identifying the figure}

\item{path}{Character string giving the folder path}
}
\value{
Prints a saved figure
}
\description{
Load and Show Saved Figures
}
\examples{

## Code
show_figure
}
