% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_missingness.R
\name{summarise_missingness}
\alias{summarise_missingness}
\title{Summarise Missingness by Variable}
\usage{
summarise_missingness(df)
}
\arguments{
\item{df}{A dataframe containing cleaned ETS data as produced by \code{tbinenglanddataclean}.}
}
\value{
A named list summarising missing data for the following variables: date of starting treatment,
date of ending treatment, date of death, and cause of death.
}
\description{
This function filters the ETS so that only entries that could have been missing are
evaluated for each variable.
}
\examples{

## Code
summarise_missingness
}
