% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_variable_missingness.R
\name{model_variable_missingness}
\alias{model_variable_missingness}
\title{Model Variable Missingness}
\usage{
model_variable_missingness(df = NULL, var = NULL, confounders = NULL,
  confounder_names = NULL)
}
\arguments{
\item{df}{A dataframe containing cleaned ETS data as produced by \code{tbinenglanddataclean}.}

\item{var}{A character string indicating the name of the variable
to explore missingness within}

\item{confounders}{A character vector containing the programmatic names of variables
to use as confounders.}

\item{confounder_names}{A character vector containing presentation names of variables. If not supplied will default
to \code{confounders}.}
}
\value{
A dataframe containing data and model estimates for the specified variable and confounders.
}
\description{
Model Variable Missingness
}
\examples{

## Code
model_variable_missingness
}
