% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_date_variable_missingness.R
\name{plot_date_variable_missingness}
\alias{plot_date_variable_missingness}
\title{Plot Missingness for a Date Variable}
\usage{
plot_date_variable_missingness(df = NULL, var = NULL,
  start_year = NULL, end_year = NULL, split_year = NULL,
  facet_by_year = TRUE)
}
\arguments{
\item{df}{A dataframe containing cleaned ETS data as produced by \code{tbinenglanddataclean}.}

\item{var}{A character string indicating the name of the date variable
to explore missingness within.}

\item{start_year}{Numeric, indicating which year to start including data from (inclusive)}

\item{end_year}{Numeric, indicating which year to stop including data from (exclusive)}

\item{split_year}{The year to use as a splitting point when facetting. By default uses the mean year.}

\item{facet_by_year}{Logical defaults to \code{TRUE}. Should the plots be facetted by a year midpoint}
}
\value{
A list of plots including: missing data by month and by day.
}
\description{
Plot Missingness for a Date Variable
}
\examples{

## Code
plot_date_variable_missingness
}
